Add dump debug data to zip file
This change adds a section in Settings > Preferences called "Debugging" with text "This will package the log files and other debugging information into a ZIP file." along with a button "Dump Debug Info" that fires a "core/dump_debug_data" IPC event to the backend to dump the collected (also a new addition in this commit) stdout and stderr output of ryod and ryo-wallet-rpc, and packages them into a ZIP file on the Desktop, or the home folder if the Desktop folder doesn't exist.

The backend sends a "debug_data_completed" IPC event back to the renderer which them displays a success modal dialog reading "Debug Data Dumped".